+++ src/lsShimInstall.ts
@@ -1,3 +1,5 @@
 import { ensureLsShimStyle } from "./lsShim";
+import { ensureThemeStyle } from "./themeGallery";
 
 ensureLsShimStyle();
+ensureThemeStyle();